convert index page to .tsx
